advisor-backend/src/answer/answer.module.ts
import { Module } from '@nestjs/common';
import { AnswerService } from './answer.service';
import { AnswerController } from './answer.controller';

// NestJS Module System helps structure application code into domain areas
//  leading to better design and improved maintainability.
// Controllers are responsible for handling incoming requests and returning responses to the client.
// Providers are responsible for executing business logic and returning data to the controller.
// Module for answer related routes
@Module({
  controllers: [AnswerController],
  providers: [AnswerService],
})
export class AnswerModule {}
